rent_pro/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.46.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-model-utils==4.2.0
django-notifications-hq==1.6.0
djangorestframework==3.12.4
idna==3.3
jsonfield==3.1.0
migrate==0.3.8
oauthlib==3.1.1
Pillow==8.3.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
stripe==2.64.0
swapper==1.3.0
urllib3==1.26.7
